+++ src/artery/application/md/kalman/KalmanFilterJ_SC.h
@@ -22,56 +22,56 @@ class KalmanFilterJ_SC {
 		int n;
 		int m; 
 
-		float A[KLM_N_SC][KLM_N_SC];
-		float AT[KLM_N_SC][KLM_N_SC]; 
+		double A[KLM_N_SC][KLM_N_SC];
+		double AT[KLM_N_SC][KLM_N_SC]; 
 
-		float B[KLM_N_SC][KLM_M_SC]; 
+		double B[KLM_N_SC][KLM_M_SC]; 
 
-		float H[KLM_N_SC][KLM_N_SC]; 
-		float HT[KLM_N_SC][KLM_N_SC]; 
-		float Q[KLM_N_SC][KLM_N_SC];
-		float R[KLM_N_SC][KLM_N_SC];
-		float I[KLM_N_SC][KLM_N_SC]; 
+		double H[KLM_N_SC][KLM_N_SC]; 
+		double HT[KLM_N_SC][KLM_N_SC]; 
+		double Q[KLM_N_SC][KLM_N_SC];
+		double R[KLM_N_SC][KLM_N_SC];
+		double I[KLM_N_SC][KLM_N_SC]; 
 
-		float X[KLM_N_SC]; 
-		float P[KLM_N_SC][KLM_N_SC]; 
-		float K[KLM_N_SC][KLM_N_SC]; 
+		double X[KLM_N_SC]; 
+		double P[KLM_N_SC][KLM_N_SC]; 
+		double K[KLM_N_SC][KLM_N_SC]; 
 
-		float X0[KLM_N_SC]; 
-		float P0[KLM_N_SC][KLM_N_SC]; 
+		double X0[KLM_N_SC]; 
+		double P0[KLM_N_SC][KLM_N_SC]; 
 
-		float Temp_1[KLM_N_SC];
-		float Temp_2[KLM_N_SC]; 
-		float Temp_3[KLM_N_SC]; 
+		double Temp_1[KLM_N_SC];
+		double Temp_2[KLM_N_SC]; 
+		double Temp_3[KLM_N_SC]; 
 
 
-		float TempN_1[KLM_N_SC][KLM_N_SC]; 
-		float TempN_2[KLM_N_SC][KLM_N_SC]; 
-		float TempN_3[KLM_N_SC][KLM_N_SC]; 
-		float TempN_4[KLM_N_SC][KLM_N_SC]; 
+		double TempN_1[KLM_N_SC][KLM_N_SC]; 
+		double TempN_2[KLM_N_SC][KLM_N_SC]; 
+		double TempN_3[KLM_N_SC][KLM_N_SC]; 
+		double TempN_4[KLM_N_SC][KLM_N_SC]; 
 
 		MatrixOp_SC matrixOp_SC = MatrixOp_SC();
 		
 
 		KalmanFilterJ_SC();
 
-		void setFixed ( float _A [][KLM_N_SC], float  _H [][KLM_N_SC], float _Q [][KLM_N_SC], float _R [][KLM_N_SC]);
+		void setFixed ( double _A [][KLM_N_SC], double  _H [][KLM_N_SC], double _Q [][KLM_N_SC], double _R [][KLM_N_SC]);
 
-		void setFixed ( float _A [][KLM_N_SC], float  _H [][KLM_N_SC], float _Q [][KLM_N_SC], float _R [][KLM_N_SC] , float _B [][KLM_M_SC]);
+		void setFixed ( double _A [][KLM_N_SC], double  _H [][KLM_N_SC], double _Q [][KLM_N_SC], double _R [][KLM_N_SC] , double _B [][KLM_M_SC]);
 
 
-		void setA(float _A [][KLM_N_SC]);
+		void setA(double _A [][KLM_N_SC]);
 
-		void setB(float _B [][KLM_M_SC]);
+		void setB(double _B [][KLM_M_SC]);
 
-		void setR(float _R [][KLM_N_SC]);
+		void setR(double _R [][KLM_N_SC]);
 
-		void setInitial( float  _X0[], float _P0[][KLM_N_SC] );
+		void setInitial( double  _X0[], double _P0[][KLM_N_SC] );
 
 		void predict ( void );
 
-		void predict( float  U []);
-		void correct ( float Z []);
+		void predict( double  U []);
+		void correct ( double Z []);
 
 };
